게시글 단건 조회 & Swagger설정

+++ src/main/java/kr/co/puerpuella/apitextssul/api/article/dto/request/Art02Request.java
@@ -4,12 +4,10 @@ import kr.co.puerpuella.apitextssul.common.framework.CommonDTO;
 import lombok.Builder;
 import lombok.Data;
 
-/**
- * ARTICLE01
- */
 @Builder
 @Data
 public class Art02Request extends CommonDTO {
 
-    private Long articleId;
+    /** 조회 대상 게시글ID */
+    private Integer articleId;
 }
